+++ src/Twigcs/Rule/Html/UnclosedVoidTagsRule.php
@@ -9,18 +9,24 @@
 
 declare(strict_types=1);
 
-namespace BitBag\CodingStandard\Twigcs\Rule;
+namespace BitBag\CodingStandard\Twigcs\Rule\Html;
 
+use BitBag\CodingStandard\Twigcs\Dto\HtmlTagDto;
 use BitBag\CodingStandard\Twigcs\Ruleset\Ruleset;
 use BitBag\CodingStandard\Twigcs\Util\HtmlUtil;
 use FriendsOfTwig\Twigcs\Rule\AbstractRule;
 use FriendsOfTwig\Twigcs\Rule\RuleInterface;
 use FriendsOfTwig\Twigcs\TwigPort\TokenStream;
 
-final class NoSpaceInAttributeRule extends AbstractRule implements RuleInterface
+final class UnclosedVoidTagsRule extends AbstractRule implements RuleInterface
 {
+    /** @var string[] */
+    private $tags = [
+        'area', 'base', 'br', 'col', 'embed', 'hr', 'img', 'input', 'link', 'meta', 'param', 'source', 'track', 'wbr'
+    ];
+
     /** @var string */
-    private $pattern = '#=\s*("[^"]*"|\'[^\']*\')(?<offset>[^\s/>])#s';
+    private $pattern = '#/\s*>$#';
 
     /** @var HtmlUtil */
     private $htmlUtil;
@@ -38,14 +44,14 @@ final class NoSpaceInAttributeRule extends AbstractRule implements RuleInterface
         $content = $this->htmlUtil->stripUnnecessaryTagsAndSavePositions($tokens->getSourceContext()->getCode());
 
         foreach ($this->htmlUtil->getParsedHtmlTags($content) as $tag) {
-            foreach ($this->getViolationNoSpaces($tag->getHtmlLine()) as $noSpace) {
-                $offset = $this->htmlUtil->getTwigcsOffset($content, $tag->getOffset() + $noSpace['offset'][1]);
+            if ($this->isTagUnclosed($tag)) {
+                $offset = $this->htmlUtil->getTwigcsOffset($content, $tag->getOffset());
 
                 $violations[] = $this->createViolation(
                     $tokens->getSourceContext()->getPath(),
                     $offset->getLine(),
                     $offset->getColumn(),
-                    sprintf(Ruleset::ERROR_NO_SPACE_BETWEEN_ATTRIBUTES, $tag->getTag())
+                    sprintf(Ruleset::ERROR_UNCLOSED_VOID_HTML_TAG, $tag->getTag())
                 );
             }
         }
@@ -53,10 +59,9 @@ final class NoSpaceInAttributeRule extends AbstractRule implements RuleInterface
         return $violations;
     }
 
-    private function getViolationNoSpaces(string $html): array
+    private function isTagUnclosed(HtmlTagDto $tag): bool
     {
-        return preg_match_all($this->pattern, $html, $noSpaces, HtmlUtil::REGEX_FLAGS)
-            ? $noSpaces
-            : [];
+        return in_array($tag->getTag(), $this->tags)
+            && !preg_match($this->pattern, $tag->getHtml());
     }
 }
