+++ src/Twigcs/Rule/Html/WhitespaceInAttributesRule.php
@@ -9,24 +9,18 @@
 
 declare(strict_types=1);
 
-namespace BitBag\CodingStandard\Twigcs\Rule;
+namespace BitBag\CodingStandard\Twigcs\Rule\Html;
 
-use BitBag\CodingStandard\Twigcs\Dto\HtmlTagDto;
 use BitBag\CodingStandard\Twigcs\Ruleset\Ruleset;
 use BitBag\CodingStandard\Twigcs\Util\HtmlUtil;
 use FriendsOfTwig\Twigcs\Rule\AbstractRule;
 use FriendsOfTwig\Twigcs\Rule\RuleInterface;
 use FriendsOfTwig\Twigcs\TwigPort\TokenStream;
 
-final class VoidTagRule extends AbstractRule implements RuleInterface
+final class WhitespaceInAttributesRule extends AbstractRule implements RuleInterface
 {
-    /** @var string[] */
-    private $tags = [
-        'area', 'base', 'br', 'col', 'embed', 'hr', 'img', 'input', 'link', 'meta', 'param', 'source', 'track', 'wbr'
-    ];
-
     /** @var string */
-    private $pattern = '#/\s*>$#';
+    private $pattern = '#=\s*("[^"]*"|\'[^\']*\')(?<offset>[^\s/>])#s';
 
     /** @var HtmlUtil */
     private $htmlUtil;
@@ -44,14 +38,14 @@ final class VoidTagRule extends AbstractRule implements RuleInterface
         $content = $this->htmlUtil->stripUnnecessaryTagsAndSavePositions($tokens->getSourceContext()->getCode());
 
         foreach ($this->htmlUtil->getParsedHtmlTags($content) as $tag) {
-            if ($this->isTagUnclosed($tag)) {
-                $offset = $this->htmlUtil->getTwigcsOffset($content, $tag->getOffset());
+            foreach ($this->getNoSpaces($tag->getHtml()) as $noSpace) {
+                $offset = $this->htmlUtil->getTwigcsOffset($content, $tag->getOffset() + $noSpace['offset'][1]);
 
                 $violations[] = $this->createViolation(
                     $tokens->getSourceContext()->getPath(),
                     $offset->getLine(),
                     $offset->getColumn(),
-                    sprintf(Ruleset::ERROR_UNCLOSED_VOID_HTML_TAG, $tag->getTag())
+                    sprintf(Ruleset::ERROR_NO_SPACE_BETWEEN_ATTRIBUTES, $tag->getTag())
                 );
             }
         }
@@ -59,9 +53,10 @@ final class VoidTagRule extends AbstractRule implements RuleInterface
         return $violations;
     }
 
-    private function isTagUnclosed(HtmlTagDto $tag): bool
+    private function getNoSpaces(string $html): array
     {
-        return in_array($tag->getTag(), $this->tags)
-            && !preg_match($this->pattern, $tag->getHtmlLine());
+        return preg_match_all($this->pattern, $html, $noSpaces, HtmlUtil::REGEX_FLAGS)
+            ? $noSpaces
+            : [];
     }
 }
